requirements.txt
amqp==5.3.1
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0

Django==5.0.4
django-jazzmin==2.6.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7

flower==2.0.1
humanize==4.9.0

idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.4
MarkupSafe==2.1.5

packaging==23.2
prometheus_client==0.19.0
prompt_toolkit==3.0.43
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1

pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4

tornado==6.4
tzdata==2024.1
tzlocal==5.2

uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1